# shared/api-contract/tts-lab-openapi.yaml
openapi: 3.0.3
info:
  title: TTS Lab API
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/audiobooks/{id}:
    get:
      tags: [Audiobooks]
      operationId: getAudiobookDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audiobook detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Audiobooks]
      operationId: updateAudiobookTitle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAudiobookTitleRequest'
      responses:
        '200':
          description: Updated audiobook detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookDetailResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/audiobooks/workflow/speaker-voice-analysis:
    post:
      tags: [Audiobook Workflow]
      operationId: analyzeSpeakerVoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakerVoiceAnalysisRequest'
      responses:
        '200':
          description: Speaker voice analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerVoiceAnalysisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitedError'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/ProviderUnavailableError'
  /api/audiobooks/workflow/speaker-split-analysis:
    post:
      tags: [Audiobook Workflow]
      operationId: splitDialogue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakerSplitAnalysisRequest'
      responses:
        '200':
          description: Speaker split analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerSplitAnalysisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/audiobooks/workflow/emotion-annotation-analysis:
    post:
      tags: [Audiobook Workflow]
      operationId: annotateEmotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmotionAnnotationAnalysisRequest'
      responses:
        '200':
          description: Emotion annotation analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/audiobooks/workflow/script-preview-save:
    post:
      tags: [Audiobook Workflow]
      operationId: saveScriptPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptPreviewSaveRequest'
      responses:
        '200':
          description: Persisted script preview turns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerSplitAnalysisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/audiobooks/workflow/final-request-preview:
    post:
      tags: [Audiobook Workflow]
      operationId: buildFinalRequestPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalTtsRequestPreviewRequest'
      responses:
        '200':
          description: Final TTS request preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalTtsRequestPreviewResponse'
  /api/audiobooks/workflow/single-speaker-render-plan:
    post:
      tags: [Audiobook Workflow]
      operationId: buildSingleSpeakerRenderPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleSpeakerRenderPlanRequest'
      responses:
        '200':
          description: Single-speaker render plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleSpeakerRenderPlanResponse'
  /api/audiobooks/workflow/create-audio:
    post:
      tags: [Audiobook Workflow]
      operationId: createAudio
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
        - name: targetSegmentIndex
          in: query
          required: true
          schema:
            type: integer
            description: Zero-based index of the speech segment to render
      responses:
        '200':
          description: Rendered audio download
          headers:
            Content-Disposition:
              schema:
                type: string
            Content-Type:
              schema:
                type: string
            X-Audiobook-Project-Id:
              schema:
                type: string
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/audiobooks/workflow/projects/{projectId}/audio-generated:
    post:
      tags: [Audiobook Workflow]
      operationId: finalizeAudiobookAudioGeneration
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current audiobook workflow snapshot after the preview is finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/audiobooks/workflow/projects/{projectId}:
    get:
      tags: [Audiobook Workflow]
      operationId: getAudiobookWorkflowSnapshot
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current audiobook workflow snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/audiobooks/workflow/projects/{projectId}/cast-approval:
    post:
      tags: [Audiobook Workflow]
      operationId: approveAudiobookCast
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated audiobook workflow snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/audiobooks/workflow/projects/{projectId}/script-approval:
    post:
      tags: [Audiobook Workflow]
      operationId: approveAudiobookScript
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated audiobook workflow snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/audiobooks/workflow/projects/{projectId}/production-settings:
    patch:
      tags: [Audiobook Workflow]
      operationId: updateAudiobookWorkflowProductionSettings
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudiobookWorkflowProductionSettingsRequest'
      responses:
        '200':
          description: Updated audiobook workflow snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudiobookWorkflowSnapshotResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  responses:
    ValidationError:
      description: Request validation failed
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    RateLimitedError:
      description: Rate limited
      headers:
        X-Request-Id:
          schema:
            type: string
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ProviderUnavailableError:
      description: Provider unavailable
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFoundError:
      description: Resource not found
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalError:
      description: Unexpected internal failure
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    AudiobookDetailResponse:
      type: object
      required: [id, title, status, speechSegmentCount, speakerCount, totalDurationSeconds, createdAt, updatedAt, speechSegments, audioAssets]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/AudiobookProjectStatus'
        speechSegmentCount:
          type: integer
          format: int32
        speakerCount:
          type: integer
          format: int32
          nullable: true
        totalDurationSeconds:
          type: integer
          format: int32
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        speechSegments:
          type: array
          items:
            $ref: '#/components/schemas/AudiobookSpeechSegmentResponse'
        audioAssets:
          type: array
          items:
            $ref: '#/components/schemas/AudioAssetResponse'
      additionalProperties: false
...
    AudiobookWorkflowProductionSettingsRequest:
      type: object
      required: [prompt, languageCode, modelName, audioEncoding]
      properties:
        prompt:
          type: string
        languageCode:
          type: string
        modelName:
          type: string
        audioEncoding:
          type: string
      additionalProperties: false
    AudiobookWorkflowSnapshotResponse:
      type: object
      required: [projectId, title, sourceLanguageCode, workflowStage, speakers, scriptTurns, annotatedTurns, productionSettings, audioAssets, audioAssetsCurrent, performanceNotesStale]
      properties:
        projectId:
          type: string
        title:
          type: string
        storyText:
          type: string
          nullable: true
        sourceLanguageCode:
          type: string
        workflowStage:
          $ref: '#/components/schemas/AudiobookWorkflowStage'
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerVoiceAnalysisItem'
        scriptTurns:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerSplitTurn'
        annotatedTurns:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSpeakerTurn'
        productionSettings:
          $ref: '#/components/schemas/AudiobookWorkflowProductionSettings'
        audioAssets:
          type: array
          items:
            $ref: '#/components/schemas/AudioAssetResponse'
        audioAssetsCurrent:
          type: boolean
        performanceNotesStale:
          type: boolean
        mergedAudioUrl:
          type: string
          nullable: true
      additionalProperties: false
...
    SpeakerVoiceAnalysisRequest:
      type: object
      required: [rawDialogue]
      properties:
        rawDialogue:
          type: string
      additionalProperties: false
    SpeakerVoiceAnalysisResponse:
      type: object
      required: [speakers, projectId, projectTitle, sourceLanguageCode, productionLanguageCode]
      properties:
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerVoiceAnalysisItem'
        projectId:
          type: string
          nullable: true
        projectTitle:
          type: string
        sourceLanguageCode:
          type: string
        productionLanguageCode:
          type: string
      additionalProperties: false
...
    SpeakerSplitAnalysisRequest:
      type: object
      required: [rawDialogue, speakers, projectId]
      properties:
        rawDialogue:
          type: string
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerVoiceAnalysisItem'
        projectId:
          type: string
      additionalProperties: false
    SpeakerSplitAnalysisResponse:
      type: object
      required: [turns]
      properties:
        turns:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerSplitTurn'
      additionalProperties: false
...
    EmotionAnnotationAnalysisRequest:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: string
      additionalProperties: false
...
    ScriptPreviewSaveRequest:
      type: object
      required: [projectId, turns]
      properties:
        projectId:
          type: string
        turns:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerSplitTurn'
      additionalProperties: false
    UpdateAudiobookTitleRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
      additionalProperties: false
    FinalTtsRequestPreviewRequest:
      type: object
      required: [prompt, speakers, annotatedTurns, languageCode, modelName, audioEncoding]
      properties:
        prompt:
          type: string
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerVoiceAnalysisItem'
        annotatedTurns:
          type: array
          items:
            $ref: '#/components/schemas/AnnotatedSpeakerTurn'
        languageCode:
          type: string
        modelName:
          type: string
        audioEncoding:
          type: string
      additionalProperties: false
    FinalTtsRequestPreviewResponse:
      type: object
      required: [input, voice, audioConfig]
      properties:
        input:
          type: object
          additionalProperties: true
        voice:
          type: object
          additionalProperties: true
        audioConfig:
          type: object
          additionalProperties: true
      additionalProperties: false
    SingleSpeakerRenderPlanRequest:
      type: object
      required: [input, voice, audioConfig]
      properties:
        input:
          type: object
          additionalProperties: true
        voice:
          type: object
          additionalProperties: true
        audioConfig:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    SingleSpeakerRenderPlanResponse:
      type: object
      required: [renderRequests]
      properties:
        renderRequests:
          type: array
          items:
            $ref: '#/components/schemas/SingleSpeakerRenderRequest'
      additionalProperties: false
...